Extract FromItem and makeFromItem into from-item.ts
`FromItem` and `makeFromItem` are in "with.ts", but it applies to more
than WITH. Move those definitions to their own file, "from-item.ts".

This is in preparation for a larger change to fix the way LEFT JOIN
works, which makes some changes to `FromItem` and `makeFromItem`. Wanted
to land this PR first in preparation.
